+++ apps/syncro/CHANGELOG.md
@@ -5,10 +5,17 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @neurodesk/webapp-components@0.2.1
+  - @neurodesk/webapp-components@0.2.2
   - @neurodesk/greedy@0.3.20260915
   - @neurodesk/synthsr@0.4.20260915
 
+
+### Patch Changes
+
+- Updated dependencies
+  - @neurodesk/webapp-components@0.2.1
+  - @neurodesk/greedy@0.3.20260915
+  - @neurodesk/synthsr@0.4.20260915
 
 ### Patch Changes